% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CircleTree}
\alias{CircleTree}
\title{Convert a hclust object to a circular dendrogram}
\usage{
CircleTree(x, labels = sprintf("#\%i", 1:length(x$order)), col,
  padding = 0.01, hang = 0.1 * max(x$height))
}
\arguments{
\item{x}{\code{\link{hclust}} object.}

\item{labels}{Chraracter vector of leaf labels.}

\item{col}{Vector of leaf colours.}

\item{padding}{Spacing between branches.}

\item{hang}{Branch lengths of the leaves.}
}
\value{
An object of class \code{CircleTree}
}
\description{
Any \code{\link{hclust}} object can be converted to a circular dendrogram
using this function and then plotted with the standard graphics functions.
}
\examples{
my.data <- iris[-5] + .5*matrix(rnorm(nrow(iris)*(ncol(iris)-1)), nrow(iris))
cl <- hclust(dist(my.data))

# Color according to class
x <- CircleTree(cl,
                labels=sprintf("\%s \%i", iris$Species, 1:150),
                col=iris$Species,
                padding=.01)
plot(x, label.col=TRUE)

# Color branches
x <- CircleTree(cl, labels=sprintf("\%s \%i", iris$Species, 1:150),
                col=cutree(cl, 3), padding=0, hang=.5)
plot(x, labels=FALSE)
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
plot.CircleTree
}

