% Generated by roxygen2 (4.0.1): do not edit by hand
\name{na.fill}
\alias{na.fill}
\title{Impute missing values with a given fill value}
\usage{
na.fill(x, fill)
}
\arguments{
\item{x}{Object.}

\item{fill}{The value that should replace missing values in \code{x}.}
}
\value{
A version of \code{x} with no missing values.
}
\description{
Impute missing values with a given fill value
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

