% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.CircleTree}
\alias{labels.CircleTree}
\alias{lines.CircleTree}
\alias{plot.CircleTree}
\alias{points.CircleTree}
\title{Plot functions for circular dendrograms}
\usage{
\method{plot}{CircleTree}(x, lines = TRUE, points = TRUE, labels,
  label.size = 1, label.col = FALSE, expand, lwd = c(1, 10), ..., bg)

\method{lines}{CircleTree}(x, lwd = c(1, 10), col, ...)

\method{labels}{CircleTree}(object, cex = par("cex"), col = FALSE, ...)

\method{points}{CircleTree}(x, col, ...)
}
\arguments{
\item{x}{\code{\link{CircleTree}} object.}

\item{lines}{Whether to plot the trees branches.}

\item{points}{Whether to plot the trees leaves.}

\item{labels}{Whether to plot the leaves' labels.}

\item{label.size}{Relative text size of labels.}

\item{label.col}{Whether to colour labels by leaf colour.}

\item{expand}{Controls how much white space should surround the plot.
Use \code{expand=1} for no white space and larger numbers for more.}

\item{lwd}{Minimum and maximum branch widths.}

\item{...}{Sent to \code{\link{plot}}.}

\item{bg}{Background colour.}

\item{col}{Colour.}

\item{object}{Same as \code{x}. Named differently to keep S3 consistency.}

\item{cex}{Relative size, see \code{\link{par}} for details.}
}
\value{
Nothing, produces a plot.
}
\description{
Class specific methods for class \code{\link{CircleTree}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
CircleTree
}

